requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.13
django-allauth==0.63.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
install==1.3.5
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.2
sqlparse==0.5.0
typing-extensions==4.12.0
urllib3==2.2.1
